pyproject.toml
[project]
name = "yun_bot"
version = "0.1.0"
description = "yun_bot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi,httpx,websockets]>=2.4.4",
    "nonebot-adapter-qq>=1.6.6",
    "nonebot-plugin-status>=0.9.0",
    "nonebot-plugin-apscheduler>=0.5.0",
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo", "single_session"]

[tool.nonebot.adapters]
nonebot-adapter-qq = [
    { name = "QQ", module_name = "nonebot.adapters.qq" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-status = ["nonebot_plugin_status"]
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
